BUG: studyId 생성 로직 변경

+++ StudyRepublic/src/main/java/org/mohajo/studyrepublic/domain/StudyIdGenerator.java
@@ -15,7 +15,7 @@ import lombok.Data;
 public class StudyIdGenerator implements Serializable {
 
 	private String prefix;
-	private int counter;
+	private double counter;
 	
 	
 }
